requirements.txt
blurhash==1.1.4
boost==0.1
certifi==2020.6.20
chardet==3.0.4
cmake==3.18.2.post1
cycler==0.10.0
decorator==4.4.2
dlib==19.21.0
future==0.18.2
idna==2.10
imageio==2.9.0
kiwisolver==1.2.0
Mastodon.py==1.5.1
matplotlib==3.3.1
networkx==2.5
nptyping==1.3.0
numpy==1.19.2
numpyencoder==0.2.0
opencv-python==4.4.0.42
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-magic==0.4.18
pytz==2020.1
PyWavelets==1.1.1
requests==2.24.0
scikit-image==0.15.0
scipy==1.5.2
six==1.15.0
SQLAlchemy==1.3.19
tifffile==2020.9.3
typish==1.7.0
urllib3==1.25.10